Add support for building with meson
Follow the OpenBMC herd and support a modern, comprehensible build
framework.

To build with meson:
  meson build
  ninja -C build
OR
  bitbake dbus-sensors

Add -Dtests to match de-facto OpenBMC meson usage conventions.

Maintain existing per-sensor config flags; however, do not define
targets when opting out of a sensor class.

One noteable omission is support for subprojects or vendoring.

The OpenBMC CI scripts look for meson.build before looking for
CMakelists so approval of this patch would change the build system
during CI to meson.

meta-phosphor change to update the recipe to meson is also coming
shortly.

+++ src/PSUSensorMain.cpp
@@ -470,7 +470,7 @@ static void createSensorsCallback(
             std::string labelPath;
 
             /* find and differentiate _max and _input to replace "label" */
-            int pos = sensorPathStr.find("_");
+            size_t pos = sensorPathStr.find("_");
             if (pos != std::string::npos)
             {
